pyproject.toml
[tool.poetry]
name            = "oracle"
version         = "0.1.0"
description     = "A project to create and manage escape rooms."
authors         = ["Rohan Sarvaria <rsarvaria@uwaterloo.ca>"]
readme          = "README.md"
packages        = [{include = "oracle"}]

[tool.poetry.dependencies]
python          = "^3.9"
"discord.py"    = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest          = "^3.0"

[build-system]
requires        = ["poetry-core"]
build-backend   = "poetry.core.masonry.api"
